issue track : #206 [STYLE] : OtherPage Ui , #207 [FEAT] : Other Page Code, #208 [FEAT] : send to email for report
