--- app/src/main/java/com/teambeme/beme/otherpage/model/Data.kt
@@ -1,9 +1,0 @@
-package com.teambeme.beme.otherpage.model
-import com.google.gson.annotations.SerializedName
-
-data class Data(
-    @SerializedName("answers")
-    val answers: List<Answer>,
-    @SerializedName("page_len")
-    val pageLen: Int
-)
